Remove repeated logic from code
