+++ src/components/SharedComponents/Typography/Subheading1.js
@@ -1,34 +1,11 @@
 // @flow
 
-import { LightText } from "components/styledComponents";
 import type { Node } from "react";
 import React from "react";
 
-type Props = {
-  children: any,
-  testID?: string,
-  style?: any,
-  className?: string
-}
+import INatTextLight from "./INatTextLight";
 
-const Subheading1 = ( {
-  children, testID, style, className
-}: Props ): Node => {
-  let textClass = "text-xl";
-
-  if ( className ) {
-    textClass = textClass.concat( " ", className );
-  }
-
-  return (
-    <LightText
-      className={textClass}
-      style={style}
-      testID={testID}
-    >
-      {children}
-    </LightText>
-  );
-};
+// eslint-disable-next-line react/jsx-props-no-spreading
+const Subheading1 = ( props: any ): Node => <INatTextLight className="text-xl" {...props} />;
 
 export default Subheading1;
